desafio-2-pipeline-documentos/pyproject.toml
[project]
name = "desafio-2-pipeline-documentos"
version = "0.1.0"
description = "Pipeline inteligente de processamento de documentos com LLM"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=2.0",
    "PyMuPDF>=1.24",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tqdm>=4.0",
]

[project.scripts]
pipeline = "src.main:main"
